fix: holiday dates updating on day of the Holiday
On Thanksgiving day I restarted Home Assistant, and my "next Thanksgiving" Wolfram Alpha query changed to Thanksgiving 2020. Looking a the sensor's history the date was correct before the restart. So, obviously restarting Home Assistant causes the sensor to update.

I attempted to correct this issue by moving the API sensor to sensor.date_next_thanksgiving and creating a template sensor.date_thanksgiving which does not use the "next_thanksgiving" date if the sensor.date_thanksgiving is currently equal to the today's date.

This attempted fix was also applied to Easter and Labor Day as well.

cc: @ccostan